privatized functions that should have been privatized from the outset
